supabase implements PGVectorClient
