+++ examples/.env.example
@@ -1,4 +1,9 @@
 OPENAI_API_KEY=ADD_YOUR_API_KEY_HERE
 SERPAPI_API_KEY=ADD_YOUR_API_KEY_HERE
 SUPABASE_PRIVATE_KEY=ADD_YOUR_API_KEY_HERE
-SUPABASE_URL=ADD_YOUR_URL_KEY_HERE
+SUPABASE_URL=ADD_YOURS_HERE
+PGVECTOR_USER=ADD_YOURS_HERE
+PGVECTOR_HOST=ADD_YOURS_HERE
+PGVECTOR_DATABASE=ADD_YOURS_HERE
+PGVECTOR_PASSWORD=ADD_YOURS_HERE
+PGVECTOR_PORT=ADD_YOURS_HERE
