+++ examples/package.json
@@ -23,17 +23,22 @@
   "dependencies": {
     "@dqbd/tiktoken": "^0.4.0",
     "@pinecone-database/pinecone": "^0.0.8",
-    "@supabase/supabase-js": "2.8.0",
+    "@supabase/supabase-js": "^2.10.0",
     "chromadb": "^1.3.0",
     "js-yaml": "^4.1.0",
     "langchain": "workspace:*",
     "openai": "^3.1.0",
+    "pg": "^8.9.0",
+    "pg-format": "^1.0.4",
+    "postgres": "^3.3.3",
     "serpapi": "^1.1.0",
     "sqlite3": "^5.1.4"
   },
   "devDependencies": {
     "@tsconfig/recommended": "^1.0.2",
     "@types/js-yaml": "^4",
+    "@types/pg": "^8",
+    "@types/pg-format": "^1",
     "@typescript-eslint/eslint-plugin": "^5.51.0",
     "@typescript-eslint/parser": "^5.51.0",
     "dotenv": "^16.0.3",
