+++ examples/src/indexes/vectorstores_pgvector.ts
@@ -1,25 +1,26 @@
 import { PGVectorStore } from "langchain/vectorstores";
 import { OpenAIEmbeddings } from "langchain/embeddings";
 import { createClient } from "@supabase/supabase-js";
+import { PGClient } from "../pgvector-supabase.js";
 
 export const run = async () => {
-  const client = createClient(
+  const supabaseClient = createClient(
     process.env.SUPABASE_URL || "",
     process.env.SUPABASE_PRIVATE_KEY || ""
   );
 
+  const pgClient = new PGClient(supabaseClient);
+
   const vectorStore = await PGVectorStore.fromTexts(
-    client,
+    pgClient,
     ["Hello world", "Bye bye", "hello nice world"],
     [{ id: 2 }, { id: 1 }, { id: 3 }],
-    new OpenAIEmbeddings(),
-    "documents",
-    "match_documents",
+    new OpenAIEmbeddings()
   );
 
   const resultOne = await vectorStore.similaritySearchWithScore(
     "Hello world",
-    .80
+    0.8
   );
   console.dir(resultOne, { depth: null });
 };
