+++ langchain/package.json
@@ -62,7 +62,6 @@
     "@dqbd/tiktoken": "^0.4.0",
     "@jest/globals": "^29.4.2",
     "@pinecone-database/pinecone": "^0.0.8",
-    "@supabase/supabase-js": "2.8.0",
     "@tsconfig/recommended": "^1.0.2",
     "@types/d3-dsv": "^3",
     "@types/pdf-parse": "^1.1.1",
