+++ langchain/src/vectorstores/index.ts
@@ -2,4 +2,4 @@ export { HNSWLib } from "./hnswlib.js";
 export { Chroma } from "./chroma.js";
 export { PineconeStore } from "./pinecone.js";
 export { VectorStore, SaveableVectorStore } from "./base.js";
-export { PGVectorStore } from "./supabase.js";
+export { PGVectorStore, PGVectorClient } from "./pgvector.js";
